refactor: clock interface into common
